// src/types.ts
export type Restaurant = {
  id: string; name: string; genre: string; address: string; rating: number; reviews: number
  price: string; distance: string; lat: number; lng: number; color: string
  video: { title: string; channel: string; views: string; age: string; thumbnail: string; videoId: string }
}